Create DefaultPhoneAccountHandle properly.
In present implementation, DefaultPhoneAccountHandle is created only
if groupId is not null, which results in voice preference resetting
to "Ask First" after reboot, irrespective of user selection, if
selected phone account doesn't have groupId.
As part of this fix, Irrespective of groupId present or not in phone
account, create DefaultPhoneAccountHandle from xml.

Change-Id: Ib59bf631487043fae1e48515f4f381ee8f0e2d94
CRs-Fixed: 1099540